+++ release/githubio/js/main_uncompressed.js
@@ -4329,7 +4329,7 @@
         changelog: [{
                 title: 'Web Version July 12 2022',
                 changes: [
-                    "Update User channel screen to use new Twitch API, the sorting now is base on the last channel you followed to show first, on the future this list may get a better sorting today the APIs don't support it",
+                    'Update User channel screen to use new Twitch API',
                     'Update Search Game to use new Twitch API',
                     'Update Search Channel to use new Twitch API',
                     'Remove Search Live no longer supported by new Twitch API',
@@ -25603,7 +25603,10 @@
         ScreenObj[key].followerChannels = [];
         ScreenObj[key].followerChannelsDone = false;
         ScreenObj[key].coloumn_id = 0;
+        ScreenObj[key].channelDataPos = 0;
+        ScreenObj[key].getFollowed = true;
         ScreenObj[key].data = null;
+        ScreenObj[key].channelData = null;
         ScreenObj[key].data_cursor = 0;
         ScreenObj[key].dataEnded = false;
         ScreenObj[key].ContentLang = Main_ContentLang;
@@ -29890,13 +29893,30 @@
                 key_pgUp: Main_usergames,
                 key_pgUpNext: Main_usergames,
                 getFollowed: true,
+                channelData: null,
+                channelDataPos: 0,
                 base_url: Main_helix_api + 'users/follows?first=' + Main_ItemsLimitMax + '&from_id=',
                 base_url_channels: Main_helix_api + 'users?',
                 set_url: function() {
                     if (this.getFollowed) {
                         this.url = this.base_url + AddUser_UsernameArray[0].id + (this.cursor ? '&after=' + this.cursor : '');
                     } else {
+                        this.channels = 'id=' + this.channelData[this.channelDataPos].to_id;
+                        var i = this.channelDataPos + 1,
+                            dataLen = this.channelData.length,
+                            len = Math.min(dataLen, i + 99);
+
+                        this.channelDataPos++;
+                        for (i; i < len; i++) {
+                            this.channels += '&id=' + this.channelData[i].to_id;
+                            this.channelDataPos++;
+                        }
+
                         this.url = this.base_url_channels + this.channels;
+
+                        if (dataLen <= i) {
+                            this.dataEnded = true;
+                        }
                     }
                 },
                 label_init: function() {
@@ -29936,34 +29956,47 @@
         ScreenObj[key].concatenate = function(responseObj) {
             if (this.getFollowed) {
                 var data = responseObj[this.object];
-
                 this.cursor = responseObj.pagination.cursor;
-                if (!this.cursor || this.cursor === '') this.dataEnded = true;
 
                 if (data.length) {
-                    this.channels = 'id=' + data[0].to_id;
-                    var i = 1,
-                        len = data.length;
-                    for (i; i < len; i++) {
-                        this.channels += '&id=' + data[i].to_id;
+                    if (!this.channelData) {
+                        this.channelData = data;
+                    } else {
+                        this.channelData.push.apply(this.channelData, responseObj[this.object]);
                     }
-
-                    this.getFollowed = false;
-                    Screens_loadDataRequest(this.screen);
-                } else {
+                } else if (!this.channelData) {
                     this.dataEnded = true;
                     this.data = [];
                     this.loadDataSuccess();
                     this.loadingData = false;
+                    return;
+                }
+
+                if (this.cursor && this.cursor !== '') {
+                    Screens_loadDataRequest(this.screen);
+                } else {
+                    //sort
+                    this.channelData.sort(function(a, b) {
+                        return a.to_login < b.to_login ? -1 : a.to_login > b.to_login ? 1 : 0;
+                    });
+                    this.getFollowed = false;
+                    Screens_loadDataRequest(this.screen);
                 }
             } else {
+                var tempData = responseObj[this.object];
+                if (tempData) {
+                    tempData.sort(function(a, b) {
+                        return a.login < b.login ? -1 : a.login > b.login ? 1 : 0;
+                    });
+                }
+
                 if (this.data) {
-                    if (responseObj[this.object]) {
-                        this.data.push.apply(this.data, responseObj[this.object]);
+                    if (tempData) {
+                        this.data.push.apply(this.data, tempData);
                         this.offset = this.data.length;
                     }
                 } else {
-                    this.data = responseObj[this.object];
+                    this.data = tempData;
                     if (this.data) {
                         this.offset = this.data.length;
                     } else this.data = [];
@@ -29972,7 +30005,6 @@
                 }
 
                 this.loadingData = false;
-                this.getFollowed = true;
             }
         };
     }